refactor(extractor): Galactic coordinates feature is no longer default
